# Changed US distances type to float, because the original version is made with it # Add mechanism in US to find eternal loops # Add router validator in globalStatistics # Add tests

+++ src/main/java/com/github/schmittjoaopedro/aco/ls/us/CoordinatesGenius.java
@@ -10,18 +10,18 @@ public class CoordinatesGenius {
 
     public Coordinate g[];
 
-    public double d[][];
+    public float d[][];
 
     public CoordinatesGenius() {
         MAX_N = 500;
         g = new Coordinate[MAX_N + 1];
-        d = new double[MAX_N + 1][MAX_N + 1];
+        d = new float[MAX_N + 1][MAX_N + 1];
     }
 
     public CoordinatesGenius(int MAX_N) {
         this.MAX_N = MAX_N;
         g = new Coordinate[MAX_N + 1];
-        d = new double[MAX_N + 1][MAX_N + 1];
+        d = new float[MAX_N + 1][MAX_N + 1];
     }
 
     public void setXYValues(int n, int cities, double x, double y) {
@@ -35,7 +35,7 @@ public class CoordinatesGenius {
         for (int i = 1; i <= (task); i++) {
             for (int j = 1; j <= (task); j++) {
                 if (i != j) {
-                    d[i][j] = graph.getEdge(i - 1, j - 1).getCost();
+                    d[i][j] = (float) graph.getEdge(i - 1, j - 1).getCost();
                 }
             }
         }
